+++ R/cbsopendata.R
@@ -1,9 +1,8 @@
 #'@importFrom utils read.csv write.csv write.table View
-#'@import httr
 #'@docType package
 NULL
 
-CBSOPENDATA = "https://opendata.cbs.nl"
+BASE_URL = "https://opendata.cbs.nl"
 API = "ODataApi/odata"
 BULK = "ODataFeed/odata"